// IntelliPMS-Frontdesk/src/app/models/room-service.model.ts
export class RoomServiceModel {
  roomServiceId: number;
  guestId: number;
  guestName: string;
  roomId: number;
  roomNumber: number;
  createdDateTime: string;
  completedDateTime: string;
  employeeId: string;
  date: any;
  time: any;
  statuss: string;
  amount: number;
  paymentStatus: string;
  waiterId: string;
  invoiceId: string;
  numOfItems: number;
  checkInId: number;
}